test(transfer): moves tests and renames file
- moves tests from `deploy.test.ts` to `transfer.test.ts`
- renames file from `transfer.test.ts` to `transactions.test.ts`

I moved the tests into a single file, because running them concurrently made the tests flakey. Specifically either the contract deployment or the first transfer test would fail because of a nonce issue.

I expect it's because both tests query `eth_getTransactionCount` and get the same result milliseconds apart. So they submit transactions with the same nonce.
